Added support for batch file decoding
Prior invocation:

    file_decoder INPUT OUTPUT

Current invocation:

    file_decoder INPUT1 INPUT2 INPUT3 [--out=FOLDER/]

Rationale: starting a process is very expensive. This cost is further
extended by Ruby interpreter startup time. Doing that for every single
file that needs to be converted is unreasonable. Thanks to being able to
pass all of the input paths at once, process needs to be spawned just
once for the whole batch operation, which means significant speedup.