Removed packing support
Recent changes made it more difficult to support archive packing. This
has finally prompted me to remove it. Keeping it requires too much
maintenance for little or no gain. The only people interested in archive
repacking would be translators or developers who help translators, and:

1. Reversing everything is impractical. There are very few translators.
2. They are better off using customized encoding tools anyway - these
   adhere to their personal workflows and can be tested much better in
   practice (by contrast, testing unpacking gives instant results).

Additionally, the codebase of this project changes quickly, and keeping
around a dead feature like that is nothing but a hindrance.

I'll keep the internal Packers around for testing purposes, but that
might change in the future.