au: fix theoretical stack overflow
For files that were recognized infinitely by some transformer despite
successful conversions, at some point the stack overflowed. The
recognition cycle of one file is now limited to 10 conversions.

In practice, this could never happen with real games, and the greatest
chance to trigger this behavior was to implement a transformer that
always recognizes every file, or that returns what it was passed.