fmt: support F&C01's MCG images (no key guessing)
Unfortunately even though the key search space is a joke (255 keys to
try), I failed to provide a generic brute force search. For every
heuristic I can think of, there's always at least one image that passes
it with wrong key and produces noise. Because of this, users need to
pass --mcg-key manually. Truly a bummer.