fmt: rename core fmt classes
Transformer --> AbstractDecoder
Archive     --> ArchiveDecoder
Converter   --> FileDecoder

Since au is not meant to reencode files nor repack archives and the
project structure isn't made for that either, it only makes sense to
call things by their true names, which is "decoder" in this case. Names
such as Converter or Archive suggested they might be also capable of
encoding stuff.

This commit doesn't fix names of existing decoders. This is going to be
dealt with in separate commits.