fmt: refactor ArchiveDecoders
This change formally breaks saving the files through FileSaver into two
stages:

    1. Reading the file table
    2. Saving the entries in the file table

The difference between this and the previous "read_table + read_file"
pattern is that up until now it was just a convention, not a
requirement. Now the file table is exposed via the public API, which
gives ArchiveDecoder users an opportunity to choose what files to save.
As a bonus, this allows to merge common implementations of read_file,
but that's something I don't want to do for now.