au: simplify collecting input file paths
To be specific, removed all the base_name busywork. At some point this
field was used to construct paths when user passed directory rather than
a file - it tried to emulate the corresponding directory subtree. But
that behavior has been altered and the files now unpack directly to CWD.
Since this code has been lying unused for a while now, I decided to let
it go.