dec/borland/tpf0: initial support
TPF0 are serialized Delphi forms. This decoder is going to be used with
QLiE PACK decoder: QLiE engine stores the game keys within the game's
icon. So far the PACK decoder looked for some magic string in the .exe
file; with this commit, it can finally read the structures properly.