scripts: remove pyutil
This script library was used for preparation of embeddable test files.
The upside of such files was that they were immutable and thus immune to
refactors and possible bugs in core classes. The downside was that they
consumed a lot of space, and in cases where they were generated with
external Python scripts (and this library), those scripts were left
unversioned.

Since 0.10, due to increasing size of the repository (46M for the
`tests/` directory) the principle has changed use hand rolled tests
written in versioned C++ for everything viable, and resort to binary
blobs only if the decoder is too complex to be covered with hand written
test cases (see: lossy `entis/mio` transformations.)

Perhaps I'll even go as far as porting old python scripts on my HDD to
proper tests, and removing existing binary files.